requirements.txt
Jinja2
markdown
pyyaml
black
ruff